add option to get all tool revisions to get-tool-list
